% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q1_spotify.R
\name{check_spotify_credentials}
\alias{check_spotify_credentials}
\title{Yetkileri kontrol eder}
\usage{
check_spotify_credentials()
}
\value{
SPOTIFY_ID ve SPOTIFY_SECRET'ı kontrol ederek true değeri ya da hata mesajı döndürür
}
\description{
Yetkileri kontrol eder
}
