% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Title}
\usage{
cluster(xy, k.nn = 10, k.cluster = 10)
}
\arguments{
\item{xy}{Data frame.}

\item{k.nn}{Number of nearest neighbours.}

\item{k.cluster}{Number of clusters.}
}
\value{
Factor.
}
\description{
Title
}
\examples{
df <- data.frame(
  x = rnorm(1000),
  y = rnorm(1000)
)
cluster(df)
}
