% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img2matrix.R
\name{ebimage_thresh}
\alias{ebimage_thresh}
\title{Convert Image to Matrix}
\usage{
ebimage_thresh(img_data)
}
\arguments{
\item{img_data}{magick image object.}
}
\value{
Integer matrix.
}
\description{
Convert Image to Matrix
}
\examples{
library(magick)

img <- system.file(package = "spatialist", "Kevin.jpeg")
img_raw <- image_read(path = img)
img_data <- image_data(img_raw)
is.matrix(ebimage_thresh(img_data))
}
