% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img2matrix.R
\name{firstLayerNotWhite}
\alias{firstLayerNotWhite}
\title{Convert Image to Matrix}
\usage{
firstLayerNotWhite(img_data)
}
\arguments{
\item{img_data}{magick image object.}
}
\value{
Integer matrix.
}
\description{
Convert Image to Matrix
}
\examples{
library(magick)

img <- system.file(package = "spatialist", "Kevin.jpg")
img_raw <- image_read(path = img)
img_data <- image_data(img_raw)
is.matrix(firstLayerNotWhite(img_data))
}
