% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Display Heatmap of Binary Matrix}
\usage{
make_heatmap(img_matrix)
}
\arguments{
\item{img_matrix}{Matrix}
}
\value{
[`Heatmap`]
}
\description{
Display Heatmap of Binary Matrix
}
\examples{
make_heatmap(matrix(sample(0:1, size = 100, replace = TRUE), nrow = 10))
}
