% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{make_jitter}
\alias{make_jitter}
\title{Jitterplot}
\usage{
make_jitter(xy_coord, jitter)
}
\arguments{
\item{xy_coord}{`data.frame` of XY coordinates.}

\item{jitter}{Amount of jitter.}
}
\value{
[`ggplot`] object.
}
\description{
Jitterplot
}
\examples{
make_jitter(data.frame(x = 1, y = 1))
}
