% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{make_point}
\alias{make_point}
\title{Scatterplot}
\usage{
make_point(xy_coord)
}
\arguments{
\item{xy_coord}{`data.frame` of XY coordinates.}
}
\value{
[`ggplot`] object.
}
\description{
Scatterplot
}
\examples{
make_point(data.frame(x = 1, y = 1))
}
