% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{make_spatial}
\alias{make_spatial}
\title{Spatial Plot}
\usage{
make_spatial(xy_coord, point.size)
}
\arguments{
\item{xy_coord}{`data.frame` of XY coordinates.}

\item{point.size}{Point size}
}
\value{
[`ggplot`] object.
}
\description{
Spatial Plot
}
\examples{
make_jitter(data.frame(x = 1, y = 1))
}
