% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatialise}
\alias{spatialise}
\title{Spatialise Images}
\usage{
spatialise(path)
}
\arguments{
\item{path}{Path to the input file.}
}
\value{
A `ggplot` object.
}
\description{
Main function.
}
\examples{
spatialise(system.file(package = "spatialist", "Kevin.jpg"))
}
