% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_spotify.R
\name{spotify_search_artist}
\alias{spotify_search_artist}
\title{Artist ID döndürür}
\usage{
spotify_search_artist(artist_name)
}
\arguments{
\item{artist_name}{Artist adı girilir}
}
\value{
Artist adına göre yakın isimdeki ID'leri döndürür
}
\description{
Artist ID döndürür
}
