% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Title}
\usage{
cluster(xy, extras)
}
\arguments{
\item{xy}{Data frame.}

\item{extras}{List of additional parameters. See Details.}
}
\value{
Factor.
}
\description{
Title
}
\examples{
df <- data.frame(
  x = rnorm(1000),
  y = rnorm(1000)
)
cluster(df)
}
