% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatialise}
\alias{spatialise}
\title{Spatialise Images}
\usage{
spatialise(
  path,
  return.type = c("raw", "flatten"),
  img2matrix.FUN = firstLayerNotWhite,
  extras = list()
)
}
\arguments{
\item{path}{Path to the input file.}

\item{return.type}{Type of image to return.}

\item{extras}{List of options for sub-functions.}
}
\value{
A `ggplot` object.
}
\description{
Main function.
}
\examples{
kevin <- spatialise(
  path = system.file(package = "spatialist", "Kevin.jpg")
)
print(kevin)

kevin <- spatialise(
  path = system.file(package = "spatialist", "Kevin.jpg"),
  return.type = "flatten",
  extras = list(
    image_flatten = list(operator = "Threshold")
  )
)
print(kevin)
}
